requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
asyncpg
greenlet
alembic
pydantic-settings
python-dotenv
